pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artifice-engine"
version = "0.1.0"
description = "A node-based glitch art co-creative environment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Artifice Engine Contributors"}
]
keywords = ["glitch", "art", "generative", "node-based", "image-processing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Graphics :: Editors",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pywavelets>=1.4.0",
    "scipy>=1.10.0",
    "PySide6>=6.5.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0.0",
    "cupy-cuda12x>=12.0.0",
]
ui = [
    "PySide6>=6.5.0",
]
video = [
    "moviepy>=1.0.3",
]
audio = [
    "librosa>=0.10.0",
    "sounddevice>=0.4.6",
]
ai = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "artifice-engine[gpu,ui,video,audio,ai,dev]",
]

[project.scripts]
artifice = "artifice.__main__:main"

[project.urls]
Homepage = "https://github.com/templeoflum/Artifice-Engine"
Repository = "https://github.com/templeoflum/Artifice-Engine"
Issues = "https://github.com/templeoflum/Artifice-Engine/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
